Add create_schemas
Started with the idea of having create_schemas and create_tables
but going to change it to create_raw_schema and create_processed_schema
so we group all related operations by schema. This gives us nice
vertical slices for each module instead of the horizontal approach
of first creating the namespaces then the tables and conflating the
types of tables for raw and processed data.

The structure of this file is designed so we can run this script
independently or part of a larger whole when calling `main.py` in which
case a connection is passed in. When run independently, we need to
load env vars and create the connection ourselves, steps the main module
would normally do for us. We only close the connection in the "if main"
block when ran independently because `main.py` closes its own connection.