Specification for storing CSV files
Initially I wanted to make a serverless event-driven project and use S3
for CSVs. But there's no time for that, so instead we're doing things
the old school way where we explicitly differentiate the stages of the
data by states: raw and processed. We'll store files in a structured
manner on our own disk, and we'll also use the DB as a staging area with
raw and processed schemas.